services/orchestrator/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
asyncpg==0.28.0
alembic==1.11.1
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.24.0
celery==5.3.1
redis==4.5.5
pytest==7.4.0
pytest-asyncio==0.22.0
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.3

